Add new create pool index
