+++ centrifuge-app/src/components/Menu/index.tsx
@@ -4,12 +4,14 @@ import {
   IconGlobe,
   IconInvestments,
   IconNft,
+  IconPlus,
   IconSwitch,
   IconWallet,
   MenuItemGroup,
   Shelf,
   Stack,
 } from '@centrifuge/fabric'
+import styled from 'styled-components'
 import { config } from '../../config'
 import { useAddress } from '../../utils/useAddress'
 import { useIsAboveBreakpoint } from '../../utils/useIsAboveBreakpoint'
@@ -22,6 +24,24 @@ import { IssuerMenu } from './IssuerMenu'
 import { NavManagementMenu } from './NavManagementMenu'
 import { PageLink } from './PageLink'
 import { PoolLink } from './PoolLink'
+
+const COLOR = '#7C8085'
+
+const StyledRouterLinkButton = styled(RouterLinkButton)`
+  width: 100%;
+  & > span {
+    background-color: ${COLOR};
+    border-color: transparent;
+    color: white;
+    margin-bottom: 20px;
+
+    &:hover {
+      box-shadow: 0px 0px 0px 3px #7c8085b3;
+      background-color: ${COLOR};
+      color: white;
+    }
+  }
+`
 
 export function Menu() {
   const pools = usePoolsThatAnyConnectedAddressHasPermissionsFor() || []
@@ -130,8 +150,8 @@ export function Menu() {
 
 function CreatePool() {
   return (
-    <RouterLinkButton to="/issuer/create-pool" small>
+    <StyledRouterLinkButton icon={<IconPlus size="iconSmall" />} to="/issuer/create-pool" small variant="inverted">
       Create pool
-    </RouterLinkButton>
+    </StyledRouterLinkButton>
   )
 }
